RDM-213 Extract validation logic from SecurityClassificationService into SecurityValidationService. Other minor cleanups.
